+++ src/index.ts
@@ -2,6 +2,7 @@ import './localization';
 import options from './options';
 import { registerServiceWorker } from './registerServiceWorker';
 import { Roulette } from './roulette';
+import { SoopController } from './soopController';
 
 registerServiceWorker();
 
@@ -9,3 +10,4 @@ const roulette = new Roulette();
 
 (window as any).roulette = roulette;
 (window as any).options = options;
+(window as any).SoopController = SoopController;